requirements.txt
Django==5.1.3
requests==2.32.3
python-decouple==3.8
django-widget-tweaks==1.5.0
dj_database_url==2.2.0
sentry-sdk==2.19.0
pandas==2.2.3
lxml==5.3.0
djangorestframework==3.15.2
requests-oauthlib==2.0.0